remove old timerun backend code
