+++ src/game/CMakeLists.txt
@@ -81,7 +81,6 @@ add_library(qagame MODULE
 	"etj_string_utilities.cpp"
 	"etj_target_init.cpp"
 	"etj_time_utilities.cpp"
-	"etj_timerun.cpp"
 	"etj_timerun_repository.cpp"
 	"etj_timerun_v2.cpp"
 	"etj_timerun_entities.cpp"
